@@ -17,22 +17,23 @@
     "start": "next start"
   },
   "dependencies": {
-    "@react-three/a11y": "^2.1.0",
-    "@react-three/drei": "^7.4.0",
+    "@react-three/drei": "^7.6.1",
     "@react-three/fiber": "^7.0.6",
-    "next": "^11.1.0",
+    "glsl-random": "^0.0.5",
+    "next": "^11.1.2",
     "react": "^18.0.0-alpha-dbe3363cc",
     "react-dom": "^17.0.2",
-    "three": "^0.131.3",
+    "short-uuid": "^4.2.0",
+    "three": "^0.132.2",
     "three-stdlib": "^2.4.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.17",
-    "@next/bundle-analyzer": "^11.1.0",
-    "autoprefixer": "^10.2.6",
+    "@babel/core": "^7.15.4",
+    "@next/bundle-analyzer": "^11.1.2",
+    "autoprefixer": "^10.3.4",
     "babel-plugin-glsl": "^1.0.0",
     "eslint": "^7.32.0",
-    "eslint-config-next": "^11.1.0",
+    "eslint-config-next": "^11.1.2",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-tailwind": "^0.2.1",
     "file-loader": "^6.2.0",
@@ -43,8 +44,8 @@
     "postcss": "^8.3.6",
     "prettier": "^2.3.2",
     "raw-loader": "^4.0.2",
-    "tailwindcss": "^2.2.7",
+    "tailwindcss": "^2.2.9",
     "url-loader": "^4.1.1",
-    "webpack": "^5.44.0"
+    "webpack": "^5.52.0"
   }
 }
